# Occassionally faster `powermod` for `BitInteger`