--- a/base/intfuncs.jl
+++ b/base/intfuncs.jl
@@ -1,4 +1,3 @@
-using Base: MultiplicativeInverses
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
 ## number-theoretic functions ##