This PR adds a fastpath to `powermod` that uses `Base.multiplicativeinverse` when the mod `m` is small enough to guarantee that the intermediate products do not overflow.
We could in theory handle the overflow cases as well, but for that someone would need to implement `mod(::W, ::MultiplicativeInverse{T})` for when `W == widen(T)`.

Due to the check, this cases a slight performance regression for the cases that don't hit the fast path.
Due to the overhead of calculating the multiplicative inverse, it causes a slight regresssion in cases where the powers are small. We could also check the power for aoviding the fast path.

A few sample benchmarks (`mypowermod` is the implementation introduced in this PR, `powermod` is `Base.powermod` in Julia 1.10):
```julia
julia> z = ((1<<41)-13212)
2199023242340

julia> w = ((1<<31)-13212)
2147470436

julia> @btime powermod($(Ref(3))[], $(Ref(10))[], $(Ref(z))[])
  50.873 ns (0 allocations: 0 bytes)
59049

julia> @btime mypowermod($(Ref(3))[], $(Ref(10))[], $(Ref(z))[])
  52.834 ns (0 allocations: 0 bytes)
59049

julia> @btime powermod($(Ref(3))[], $(Ref(10))[], $(Ref(w))[])
  49.930 ns (0 allocations: 0 bytes)
59049

julia> @btime mypowermod($(Ref(3))[], $(Ref(10))[], $(Ref(w))[])
  62.796 ns (0 allocations: 0 bytes)
59049

julia> @btime powermod($(Ref(3))[], $(Ref(100))[], $(Ref(z))[])
  89.881 ns (0 allocations: 0 bytes)
733612482261

julia> @btime mypowermod($(Ref(3))[], $(Ref(100))[], $(Ref(z))[])
  92.701 ns (0 allocations: 0 bytes)
733612482261

julia> @btime powermod($(Ref(3))[], $(Ref(100))[], $(Ref(w))[])
  88.981 ns (0 allocations: 0 bytes)
924062981

julia> @btime mypowermod($(Ref(3))[], $(Ref(100))[], $(Ref(w))[])
  76.081 ns (0 allocations: 0 bytes)
924062981

julia> @btime powermod($(Ref(3))[], $(Ref(1000))[], $(Ref(w))[])
  147.691 ns (0 allocations: 0 bytes)
1731397673

julia> @btime mypowermod($(Ref(3))[], $(Ref(1000))[], $(Ref(w))[])
  96.322 ns (0 allocations: 0 bytes)
1731397673

julia> @btime powermod($(Ref(3))[], $(Ref(10000))[], $(Ref(w))[])
  177.034 ns (0 allocations: 0 bytes)
258886353

julia> @btime mypowermod($(Ref(3))[], $(Ref(10000))[], $(Ref(w))[])
  105.993 ns (0 allocations: 0 bytes)
258886353
```