@chriselrod 2024-08-05
Maybe? Or increases complexity, increasing performance in some cases. 
Maybe someone could do better with a little thought.