@adienes 2025-09-17
I built off this a bit and I think it is indeed possible to get better performance in basically all cases. you can see a messy draft https://github.com/adienes/julia/pull/7  (which also catches one or two edge cases this PR didn't yet)

but the test cases I added there are failing due to a combination of https://github.com/JuliaLang/julia/issues/58025 and https://github.com/JuliaLang/julia/issues/59579